Rename ArcGisMapServerCatalogGroup to ArcGisCatalogGroup.
