+++ lib/Models/registerCatalogMembers.js
@@ -2,7 +2,7 @@
 
 /*global require*/
 
-var ArcGisMapServerCatalogGroup = require('./ArcGisMapServerCatalogGroup');
+var ArcGisCatalogGroup = require('./ArcGisCatalogGroup');
 var ArcGisMapServerCatalogItem = require('./ArcGisMapServerCatalogItem');
 var CkanCatalogGroup = require('./CkanCatalogGroup');
 var createCatalogMemberFromType = require('./createCatalogMemberFromType');
@@ -24,8 +24,8 @@ var registerCatalogMembers = function() {
     createCatalogMemberFromType.register('csv', CsvCatalogItem);
     createCatalogMemberFromType.register('czml', CzmlCatalogItem);
     createCatalogMemberFromType.register('esri-mapServer', ArcGisMapServerCatalogItem);
-    createCatalogMemberFromType.register('esri-mapServer-group', ArcGisMapServerCatalogGroup);
-    createCatalogMemberFromType.register('esri-group', ArcGisMapServerCatalogGroup);
+    createCatalogMemberFromType.register('esri-mapServer-group', ArcGisCatalogGroup);
+    createCatalogMemberFromType.register('esri-group', ArcGisCatalogGroup);
     createCatalogMemberFromType.register('geojson', GeoJsonCatalogItem);
     createCatalogMemberFromType.register('gpx', GpxCatalogItem);
     createCatalogMemberFromType.register('group', CatalogGroup);
