+++ lib/ViewModels/AddDataPanelViewModel.js
@@ -1,7 +1,7 @@
 'use strict';
 
 /*global require,ga*/
-var ArcGisMapServerCatalogGroup = require('../Models/ArcGisMapServerCatalogGroup');
+var ArcGisCatalogGroup = require('../Models/ArcGisCatalogGroup');
 var addUserCatalogMember = require('../Models/addUserCatalogMember');
 var createCatalogItemFromFileOrUrl = require('../Models/createCatalogItemFromFileOrUrl');
 var loadView = require('../Core/loadView');
@@ -194,7 +194,7 @@ function loadWfs(viewModel) {
 }
 
 function loadMapServer(viewModel) {
-    var mapServer = new ArcGisMapServerCatalogGroup(viewModel.terria);
+    var mapServer = new ArcGisCatalogGroup(viewModel.terria);
     mapServer.name = viewModel.url;
     mapServer.url = viewModel.url;
